requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
constants==0.6.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
outcome==1.3.0.post0
packaging==23.2
pip-search==0.0.12
pluggy==1.4.0
Pygments==2.17.2
PySocks==1.7.1
pytest==8.0.0
pytest-html==4.1.1
pytest-metadata==3.1.0
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
